feat: Application v2.0 done.

+++ src/services/deleteRegister.service.ts
@@ -1,10 +1,17 @@
+import { response } from "express";
 import AppDataSource from "../data-source";
 import { Registers } from "../entities/registers.entities";
+import { IUserAuth } from "../interfaces/user.interfaces";
 
-const deleteRegisterService = async (iid) => {
+const deleteRegisterService = async (iid: any, user: IUserAuth) => {
     const registerRepository = AppDataSource.getRepository(Registers)
 
     const register = await registerRepository.findOneBy({ id: iid })
+
+    if (register.userId !== user.id) {
+        return 401
+    }
+
     await registerRepository.remove(register)
 
     return 204
